app-plane/services/backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.0
social-auth-app-django==5.4.0
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2

# Password Hashing
argon2-cffi==23.1.0

# Temporal Workflows
temporalio==1.5.1

# AI
anthropic==0.18.1

# Storage (S3/MinIO)
boto3==1.34.34
botocore==1.34.34

# Search
meilisearch==0.31.0

# Payments
stripe==8.2.0

# Observability
prometheus-client==0.19.0
django-prometheus==2.3.1
python-json-logger==2.0.7
python-logging-loki==0.3.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-django==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-exporter-otlp==1.22.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
openpyxl==3.1.2
pandas==2.1.4
python-magic==0.4.27
redis==5.0.1
docker==7.0.0  # Docker SDK for container management

# Email
sendgrid==6.11.0

# Development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-asyncio==0.23.3
factory-boy==3.3.0
faker==22.0.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
django-stubs==4.2.7

# Production
gunicorn==21.2.0
uvicorn==0.25.0
whitenoise==6.6.0

# Jaeger Tracing
jaeger-client==4.8.0
opentelemetry-exporter-jaeger==1.21.0

# Playwright E2E Testing
playwright==1.41.1
pytest-playwright==0.4.4
pytest-asyncio==0.23.3

# Test Utilities
faker==22.0.0
factory-boy==3.3.0
